errorpages: Restyle 503/php-fatal error to match Varnish error
The Varnish error page was originally based on the 503/php-fatal page,
including all the localisations, "try again", Freenode, donate,
and "please include the details below." components.

This was updated in Puppet 2015 to match the redesigned "Unknown domain"
and "404.html" page (commit a806dd0a61f0afb9c5).

It kept only the "try again" and "please include the details below." parts.

I don't know how often the 503.html is used (as far as I can tell, Varnish
will capture that in most cases and display its own page). But let's update
it and the php-fatal pages to match the ones from Varnish.

This commit also updates 404.html/404.php/default.html to feature an
element for `<div class="content" role="main">`, matching the updates that
were made to the Varnish error page. (Visually this is a no-op.)

Bug: T113114
Change-Id: I15295835a1a1dc6b703e2c7c90e398eaf5840c53