Update patch set 1
Patch Set 1:

If this is what we want, then ok, but I think the problem with the query is the 'all' option.

I think the query might be okay enough when one (and only one, including item) of the entity types is selected. If you / hoo / Jaime think it's still a problem for items, then okay w/ me to set this.

Patch-set: 1