Update patch set 2
Patch Set 2:

If we want this to work in beta, we need first to merge my latest confctl changes, build a package and then populate the beta etcd instance with data. I could manually populate that instance, but it would mean the test will not be complete.

Patch-set: 2
Label: Verified=0