wmgUseGettingStarted false for dewiki
Over the next 10 days WMDE will be running a
banner campaign with it's own guided tour to
be shown to newly registered users on dewiki.

GettingStarted must be disabled on dewiki to
avoid 2 tours being shown to users at the same
time.

Once this campaign has ended GettingStarted
could be swithced back on.

Change-Id: I244ab259a7278b2a971eaec4dfa733d40aaa370f