Update patch set 5
Patch Set 5: Code-Review-1

Dereckson: Incorrect - these are in fact where the current files came from. Some are from the standard PNG rendering on Commons (almost all language variants have a dedicated PNG one on Commons for the 135px size), others are rendered by Commons from the SVG indeed - this is only done typically for the larger sizes because the text quality suffers on the default SVG renderings - which is counteracted when it is used for HiDPI and essentially downscaled again.

Anyway, the purpose of this script is to 1) document where the current files came from without changing anything, 2) make it easy to repopulate.

Patch-set: 5
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1