Configure multimedia search template boosting
Multimedia results on commons generally seem a bit better than
multiedia search from other wikis on commons data. This looks to
be because commons is used template boosting.

This may have the undesired effect of making it slightly harder to
find wiki-local content, but that effect is expected to be fairly
minimal. Additionally there is a search keyword 'local:' which allows
removing commons results from multimedia search.

Bug: T163223
Change-Id: I97e368c2d73330c7fb99777fb81d3075ea30a290