Update patch set 3
Patch Set 3: Code-Review-2

I think the whole idea here is to avoid spitting an error in the logs, but still sending an error message to the user.

This is why die() is used instead of throwing an exception.

So I'd say the correct behaviour as far as we're concerned is what we have right now:

die() in web context just prints the message and exits from the web request.

A better, cleaner approach would be to return the proper http status to the client (401 or whatever else is appropriate) and then return.

Patch-set: 3
Reviewer: Giuseppe Lavagetto <1616@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2