Update patch set 1
Patch Set 1:

If an individual test changes globals we can enable it there with "@backupGlobals enabled" in that test class. But I don't think we should enable it by default. It has an impact on test execution speed and is also discouraged in PHPUnit 6.0 and higher. Before that it was enabled by default. But is disabled starting in PHPUnit 6.

Patch-set: 1
Label: Code-Review=0