Update patch set 1
Patch Set 1:

Indeed the run time goes from 5 seconds to 1minute 10 on my machine.  That is most probably due to the global space being filled via bootstrap.php which loads multiversion etc.

It is very easy for one to not explicitly enable globals backup, which then leak to other tests and that is annoying. I will try to optimize it  and if I can't find a solution will switch to @backupGlobals enabled.

Patch-set: 1