db-codfw.php: Depool db2043, db2061
db2043 and db2061 need to get one of their power cables changed. While
this should not impact the host, I would prefer to stop replication and
mysql just in case they get their power down for any unexpected reason.

db2043 is now serving dumps and vslow, but it is inactive. So pool
db2057 in for that service just to have the file organized

db2061 is API, so pool in db2054 which is the (inactive) dump, vslow
server for that shard

Bug: T163339
Change-Id: Iec2172db88702e820d65365dec67895290ac801f