Update patch set 1
Patch Set 1:

I noticed similar problems with that aggressive SVG compressor before: https://phabricator.wikimedia.org/F7748897

Also I wonder. I assume this SVG is the "stack of discs" icon. But isn't this the patch for the Wikimedia cluster configuration? From what I understood we wanted the disc icon to be part of the Wikibase software, but the Wikidata logo for the configuration on our cluster.

Patch-set: 1