Update patch set 1
Patch Set 1:

It seems I'm confused too but to my understanding what Lydia meant was that we use this logo (with Wikidata colors) for Wikimedia installations and we make another one with other colors for third parties but when I read it again, both can be interpreted :/ Is Lydia around to clarify? I think she won't be in the office today.

Patch-set: 1
Label: Verified=0