Update patch set 2
Patch Set 2:

What I see is a very much distorted version of the Wikidata logo, enforced to be a square (but the logo is not square), and also slightly misplaced in the viewBox. Not much is left of what I originally proposed.

I will do a version and upload it on Phabricator.

Patch-set: 2