Update patch set 2
Patch Set 2: Code-Review+1

DBA concern was a lack of documentation in the incident report, concern now fixed. The team asserts the issue has been triggered by the DC switchover (symptom), with an unknown cause. Kartik offered to disable it preventively before and reenable if after in the switchover procedure. The need for monitoring has also been ack'ed in the report. DBA concerns are so addressed and we can proceed with this, as soon as someone from Language engineering requests deployment.

Patch-set: 2
Reviewer: Dereckson <287@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1