scap clean: Use list to delete, rather than list to keep
Having a list to keep is making the deletion command become increasingly
complicated as we find further patterns we want to retain.

Instead, just have a list of crap we want gone. That's the most important
list anyway. Can always add more as we find things we can nuke

Adjusts the regex a tad so we can support files that don't have a suffix

Change-Id: I5d6b0f54fd5d5807b8da2a247aca3dfed6cbbea4