Update patch set 9
Patch Set 9:

(1 comment)

Failing fast is fine by me. It makes things simpler since all we have to do is call EtcdConfig::get(), and it will throw a ConfigException if the key is missing, or if anything else went wrong. Edited accordingly.

As I said on phabricator, simplicity is important, I don't think having snapshots on disk is the right way to go.

Patch-set: 9