Update patch set 9
Patch Set 9:

The difficulty is that this was meant to be used in production for the codfw->eqiad switch scheduled for May 3. Other options for implementing prefix fallbacks were discussed on #mediawiki_security on April 20 and the result of that discussion is what you have here. I'd be happy to refactor it later, but the goal of this change is for a pilot deployment to be used on May 3.

Patch-set: 9
Label: Verified=0