Update patch set 11
Patch Set 11: Code-Review+1

(1 comment)

(one nit). @Tim: Checked the logs. It did mention the two connections side-effect, but I'm not sure if that was still obvious at the point that "#5: Use MultiConfig" was mentioned. Anyhow, as long as the failure scenario is addressed I don't mind the two fetches if we agree to re-evaluate after the switchback and try to consolidate them in a way that satisfies both arguments (1: Call Config in a way that is compatible with mediawiki core for ReadOnly, e.g. not etdc-specific slashes, 2: use a cache key, presumably by using a single EtcdConfig instance only).

Patch-set: 11
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1