Remove redundant Dashiki config
Now that the Dashiki extension.json is fixed and deployed, we can remove
the redundant config from CommonSettings.php.  This is hopefully the
permanent fix for the configuration misunderstanding between Dashiki and
mediawiki-config.  For reference, this series of changes helped cause
the problem in the first place:

https://gerrit.wikimedia.org/r/#/c/344007/
https://gerrit.wikimedia.org/r/#/c/344209/
https://gerrit.wikimedia.org/r/#/c/346661/

And there was another failed attempt to deploy this patch already:

https://gerrit.wikimedia.org/r/#/c/348625/

Change-Id: I2c9d064db5ca138b99bb452ed714398b502d1440