Set oresDamagingPref default to values that actually exist
Also refactor to break this setting out into InitialiseSettings.php.

Note that this changes the wikidata setting from 'soft' to 'hard':
the old code had an if statement specially for wikidata setting it to 'soft',
but the new code has no special code for wikidata so it uses the default,
which is 'hard'. This is needed because wikidata has no likelybad
but it does have maybebad.

Bug: T165011
Change-Id: I2a9517689fc616a8c930d8758492108fdd06e9b0