Update patch set 1
Patch Set 1:

I disagree. Any remaining patent risk is about as big for MP3, as for Ogg vorbis and other free A/V codecs that we use (although for VP google is promising to take on all usage risk for anyone).

Unregistered patent holders can always turn up for any software (in the US).

Patch-set: 1