Update patch set 1
Patch Set 1:

(2 comments)

That is a good first pass!!

I would like it to avoid copy pasting code and factor out the common code somehow :]  Maybe as a static function that is shared by both code?

setUp() is run on every tests, and this suite is probably going to have a lot of test. So the code to load wgConf from InitializeSettings.php should run only once. That can be done in a follow up change though.

Patch-set: 1