Allow absolute script path for getMediaWikiCli()
This patch makes mwscript to use absolute path if given.
This enables to run scripts that are not in mediawiki repo via
mwscript.

This does not add security concerns since a) it's command-line only
and b) you could always do it with adding ../../../.. etc. but
with this patch you can do it with just providing absolute path.

Change-Id: I1a72aab4b2fa0580ae3603cbf4e18c8da4f66c6c