Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

Oops, we can actually delete most of this. That db server is gone, and nothing outside the payments cluster can talk to its replacement.

So the PrivateSettings example change can be omitted, and everything but the 'FundraiserMaintenance' flags can be deleted from CommonSettings

Patch-set: 6
Label: Code-Review=-1