test: be strict regarding globals
PHPUnit is able to detect tests mangling globals (--strict-global-state)
Set beStrictAboutChangesToGlobalState in phpunit.xml.  Since that only
act on tests, we still need WgConfTestCase::setGlobals().

[tests/WgConfTestCase.php]

When a global had a null value, we failed to unset it from the global
scope.  Port the logic from MediaWikiTestCase so we keep track of null
globals and properly unset() them.

Enhance exception message when globals are left behind.

[tests/dbconfigTest.php]

MWReal.php set $wmfRealm and $wmfDatacenter which we have to save and
restore.  $wmfDatacenter however is not set but the method left it set
to a null value.
Stop treating $wmfDatacenter has a global variable and thus stop
saving/restoring it.
Later we can port that suite to use WgConfTestCase but I have hit a wall
trying to do it.

[tests/noc-conf/highlightTest.php]

Make sure super global $_GET is unset after use.

Change-Id: Ib7a1e7cb0dfe7c9481e850a6a6fc89c91de18faf