test: be strict regarding globals
PHPUnit is able to detect tests mangling globals (--strict-global-state)
Set beStrictAboutChangesToGlobalState in phpunit.xml.  Since that only
act on tests, we still need WgConfTestCase::setGlobals().

[tests/WgConfTestCase.php]

When a global had a null value, we failed to unset it from the global
scope.  Port the logic from MediaWikiTestCase so we keep track of null
globals and properly unset() them.

Enhance exception message when globals are left behind.

[tests/dbconfigTest.php]

Port it to WgConfTestCase to get rid of the custom globals save/restore.
The sole globals required are $wmfRealm and $wmfDatacenter for
getRealmSpecificFilename().
Move $wmfDatacenter to local scope.

[tests/noc-conf/highlightTest.php]

Make sure super global $_GET is unset after use.

Change-Id: Ib7a1e7cb0dfe7c9481e850a6a6fc89c91de18faf