db-readonly: Change the read only message for something generic
Codfw is currently a passive datacenter and it should not be
publicly reachable. Removing references to a failover.

Not giving explicitly a way to report the issue because it
traditionally puts the irc network or the phabricator site down,
but at least not referencing the dc failover. Also, these should
not be shown right now to anyone but internal monitoring.

This is scheduled to be moved to etcd soon, so not wasting much
time on it.

Bug: T166345
Change-Id: I861fdc9ab7dca93f46e0ace5e328919628f91ee1