Update pre-2010 high-density Wikipedia logos
This commit updates high-density logos for the Pennsylvania
German, Polish and Venetian Wikipedias from the pre-2010
versions to the versions that are currently in use.

In case of the Venetian Wikipedia, the regular (normal-density)
logo is also being update to match the Foundation's visual
identity guidelines (particularly the font used for the subline,
ie. "The Free Encyclopedia").

Change-Id: Iae60e2be32ef1e2a2fcd403f260a86658db76c81