parsercache: Switchover pc1004 and pc2004 to db1096 and db2072
In order to solve parsercache issue, failover the first of the
server pairs to some large servers. While these have been
replicating for some hours from the original servers, they are
mostly empty, which can lead to some performance issues.

However, this would be the same effect if 1 server would fail-
the other 2 should be able to substain most of the traffic. Let's
try failing one and see the load it takes- if it is too much, the
HA system is broken anyway.

Bug: T167567
Change-Id: I2b3c0c7ced71291bf77388a888deb2f34027e78a