Update patch set 1
Patch Set 1:

Once we have it for all wikis, it should be fairly simple to drop the project-specific ones and just default to /dbname.png.

The current /static urls is at least one step forward in the right direction. Before this, it wasn't $stdLogo for most wikis. Before this, we had an almost equal number of absurd and hardcoded urls. The only difference being that those were urls for files on Commons.

Local uploads for the logo hasn't been common for quite a while. Only a handful of wikis still did that, before we moved them to static/.

See T98640 for more info.

Patch-set: 1