Update patch set 12
Patch Set 12:

fixme: This broke most usage of wmf.png. The symlinks for project-logos are fine, since those are used by CSS where the change from square to rectangle doesn't matter. (The render box for the logo is already reserved either way.)

The problem is direct usage of wmf.png, for example on the 404 page, and many many other places. https://en.wikipedia.org/404

I'd recommend creating a duplicate of wmf.png for use by logo symlinks, and then replacing wmf.png with a square variant (probaly just add some spacing). Then we can statically find usage of wmf.png and replace it with the new file + dimensions.

Re-using the same file name is problematic because it's impossible to statically determine whether it's broken or not.

Patch-set: 12