static/images: Restore dimensions of wmf.png
Follows-up 94703bf5ab3902b267d5942e9c871d05b7a64305.

Before that commit, wmf.png was square (135x135) and used in two ways:
* As symlink from project-logos for a wiki's logo.
* Directly for use in static HTML content (e.g. 404 error page for mediawiki,
  for Varnish, for Continuous Integration, and many other places).

In that commit, the dimensions changed to 135x101. That is fine for site logos
(which are auto-centered and have size computed in CSS).

However for static usage, the dimenions must be known in advance and are
currently stretched to 135x135.

This commit:
* Copy wmf.png to wmf-logo.png and update symlinks.
* Create wmf-square.png which restores the previous dimenions.
* Replace wmf.png with symlink to wmf-square.png.

Going forward, we can statically audit usage of wmf.png and replace
either with wmf-square (if a square is required) or replace with wmf-logo
and update the dimensions at the same time.

> convert wmf.png -gravity Center -background Transparent -extent 135x135 wmf-square.png
> optipng -o9 wmf-square.png
> ls -sf wmf-square.png wmf.png

Bug: T144254
Change-Id: I737e6f9fced18a44bbd4881c5294392001fda48f