Disable wgEnableJavaScriptTest on test2wiki
When I added this special page in core in 2012, I asked Reedy
to enable it (a95653477). However, it isn't meant to be used
in prod and never worked there due to various configuration
conflicts and extra test-data directories not being exposed.

If we actually want to run unit tests in production, we can
consider enabling it on Beta, but that's a separate request.

Also remove config for wgIncludejQueryMigrate which hasn't
existed for a while now in MediaWiki.

Change-Id: I8ce28a4ce79eb5bb3e2a10bac14a266204b959a5