Enable WikibaseQualityConstraints statements
This setting of the WikibaseQualityConstraints extension enables a hook
that runs on each Wikibase change, and which (if the change is on an
entity of type “property”) schedules a job (“constraintsTableUpdate” job
command) which imports constraint statements on the entity into the
wbqc_constraints database table.

This allows Wikidata editors to edit constraints immediately on the
property, instead of editing the constraint template on the property
talk page and then having to wait until the next time that the templates
are imported using the UpdateConstraintsTable.php maintenance script of
the WikibaseQualityConstraints extension (in the past, this was
generally done once every few weeks or months).

Bug: T169647
Change-Id: I8201bef3fd87793944a95af7686e53e4bd1a41a9