Alter ContentTranslation default namespace destination for zhwiki
Chinese Wikipedia requests that the default namespace where
ContentTranslation publishes is their Draft: namespace (NS #118).

This patch therefore adds 'wmgContentTranslationTargetNamespace' to
InitialiseSettings.php (which was not defined previously), sets zhwiki
to publish in NS 118 and every other wiki where CX is enabled to
NS_MAIN, which is the default.

Needs review from the Language Engineering team.
Bug: T168727

Change-Id: Iff3cfd5b161c9ee5e31228f4f76f8fde57e88965