Update patch set 1
Patch Set 1: Code-Review-1

I think the feature title (codemirror-beta-title) fails to communicate to users what it is doing, and is confusing as we already have syntax highlighting. I'd suggest calling this "Wikitext syntax highlighting" or "Syntax highlighting as you edit" or something similar.

Also the "screenshot" image is rendering wrongly, out of line for me. Not sure why.

Otherwise the beta feature package is great.

I know it's "only" testwiki, but this is the only gate for Beta Feature quality, so...

Patch-set: 1