Remove compact language links dblist for simplicity (no-op)
This retains the current situation, whilst making it simpler and
more convenient to get it enabled by default on new Wikipedias.

This change removes the compact-language-links dblist and all
references to it. All of the current list consists of Wikipedias,
and there are only two Wikipedias in which the feature is still
in beta, so it's easier to list them in InitialiseSettings.php.

Change-Id: Ifb241eb13a0af040c1ba8ed323b1c0709edcd509