[WIP] Make compact language links default for all Wikipedias except en and de
This is the situation already. This change makes it
more convenient to get it enabled by default on
new Wikipedia editions.

This change removes the compact-language-links dblist
and all references to it.
All of the current list consists of Wikipedias, and there
are only two Wikipedias in which the feature is still in beta,
so it's easier to list them in InitialiseSettings.php.

Change-Id: Ifb241eb13a0af040c1ba8ed323b1c0709edcd509