Add P279 to $wgPropertySuggesterClassifyingPropertyIds
The data produced by the PropertySuggester analyzer script treats
P279 as a classifying property. This change makes the configuration
of the PropertySuggester extension consistent with the data.

This should be configured a bug fix, since it ensures that the data
is interpreted consistently.

DEPLOY: confirm that property suggestion works properly on items with
P279 (subclass-of) set.

Bug: T169060
Change-Id: I30630c75893149c7c784e80733429975aeeafe87