Update patch set 2
Abandoned

This is not completely supported from upstream https://github.com/facebook/hhvm/issues/7854 and I didn't get the feeling that upstream is interested in testing/working on this code. I did some tests and didn't manage to obtain any good result, there might be more bugs to solve. This patch is probably a blackhole of time and imho probably not worth to follow when we could use more stable tools like nutcracker.

Patch-set: 2
Status: abandoned