Allow bureaucrats on WMF wikis to grant and remove 'confirmed'
There was some consensus to make this avalaible for administrators but
there also was some concerns as well as if this was really needed for small
projects who do not care about user rights management. As such, and
given the current Meta-Wiki policy on not granting 'bureaucrat'
permissions but to well stablished communities that demonstrates a need
of one, plus other series of securities to avoid wiki takeover episodes
as we've seen in the past; I think granting this to bureaucrats will
satisfy both parties. This is without prejudice of individual wikis
willing their administrators to be able to grant and remove this
permission as well and without prejudice of leaving the current
configuration untouched as well: that is, wikis that have already
granted this to administrators or bureaucrats won't have their
configuration changed now unless it becomes redundant in which case a
proposal for cleanup would have to be filed and actioned.

Bug: T101983
Change-Id: Ia62ad0f40e8b71b7fa424b94da48feebaa5ebfa0