Update CirrusSearch AB test rescore profiles
Setting the query_weight of the rescore profile to 0 causes pagination
to be non deterministic, because we assign 0 to everything. Instead of
assigning 0 give the rescore query a huge boost, and leave the original
scores as they were.

This needs I4d1bfc8d818 to be deployed first so ltr rescore is always
positive.

Bug: T171212
Change-Id: I85a881bada67b4cb776f8183aab8a3ca2fb7b9da