Update patch set 1
Patch Set 1:

What is the default for $wgRCFeeds from wmf-config without this override? Is that suitable for labs? Looks like it might wrongly inherit a prod thing. Looks like we should disable wmgUseRC2UDP in beta and/or reset to empty array. (also, the IRC-RC should probably be in *Services.php and maybe have a labs equivalent)

Patch-set: 1
Label: Code-Review=0