mariadb: Adding rack allocations, some formatting fixes, read-only
Adding row and rack numbers to all hosts, not only the metadata ones.

Codfw is currently a passive datacenter and it should not be publicly
reachable. Removing references to a failover. Not giving explicitly
a way to report the issue because it traditionally puts the irc network
or the phabricator site down, but at least not referencing the dc
failover. Also, these should not be shown right now to anyone but
internal monitoring.

This is scheduled to be moved to etcd soon, so not wasting much time
on it.

Also apparently, db1051 is in B3, according to racktables.

Bug: T172459
Change-Id: I9eee9a9cebbeb94072361dff73c72e7c8759ab27