Update patch set 2
Patch Set 2:

> why db1028 would serve more traffic if it has poorer HW?

They key is the title: pool db1069 with low  weight- it is a new host, so it is pooled with traffic, but with a small amount of it, in case something like the ip is wrong, or there is any problem. Once no errors or performance issues is checked, I will pool them 50%-50%, and finally, I will depool db1028- basically, a gradual weight given for this role most of the data will be cold.

Patch-set: 2
Label: Verified=0