Remove Q25267 (degree Celsius) from conversion config.
degC is not a SI unit, it is derived from actual system unit (kelvin).
However, we do not support non-multiplicative unit conversions for now,
so we can not normalize such values correctly. We should not pretend
normalizing degC values to themselves is a correct normalization.
Correct normalization would be to convert to kelvins.

Bug: T174353
Change-Id: If3c68317f649adcd62ad3deb7a51ef676ef4c4ec