copy squid.php->reverse-proxy.php, squid-labs->reverse-proxy-staging
As the first step of safely renaming these files and stop using the
"squid" term for them, simply copy the existing files.

Also stop using the word "labs". So:

cp squid.php reverse-proxy.php
cp squid-labs.php reverse-proxy-staging.php

And then we are adding the new files to createTxtFileSymlinks.sh
without removing the old ones.

This is a (partial) new version of Ibc464a1f679143912df.

Was hard to rebase and needs to be done in 2 steps.

This is still wanted because it's the last blocker to also
close T131902.

Bug: T104148
Change-Id: Iffd335598d58d0444521107123800557c126c8e3