Update patch set 4
Patch Set 4:

Um, I'm not sure I like this pattern. We don't include things like this into InitialiseSettings. And considering the size of the included file, it's going to massively explode the size of the cached IS file.

Why couldn't we wrap it in some config variable and just include it conditionally in CommonSettings?

Patch-set: 4