Update patch set 3
Patch Set 3:

> Just to check, was this done intentionally?

No, I don't think so. This is what we normally do when an extension is installed on some wikis but not in all of them, as we did with FlaggedRevs for example, or Flow. There are examples of that in the CommonSettings.php file.

Patch-set: 3
Reviewer: MarcoAurelio <787@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0