Update patch set 2
Patch Set 2:

There's no on-wiki discussion nor Phabricator ticket for this. As far as I remember this all came on #wikimedia-operations so they could promote someone to accountcreator, but they noticed later that they forgot to assign bureaucrats the ability to remove the permission as well. To remove the group now you'll need a steward. This fixes an oversight from the first patch. I'll create a Phab task though, although I'm not the owner of this one.

Patch-set: 2