Switch CirrusSearch MLR model for enwiki to older model
The model we enabled for full production usage has not shown the
same stats changes as in our AB test. Since our AB test used a slightly
different version of the model, lets switch over to that for a few
days to see if the difference is in the model itself rather than some
other, unexplained, phenomenon.

Change-Id: Idd6004625fe66cb54c62ab1164c95f8192a685a1