Update patch set 3
Patch Set 3:

I agree with Pmiazga, this should be permission based. I don't think that it should use the "staff" group specificly (Since that group is like wiki "root") but its easy to create random groups.

Additionally, what's the point of hiding this? Is it a security risk for normal users to find it (In which case, I'd like to know details). If its not a security risk, what really is the problem with having a temporary undocumented hard to discover test interface that is open to anyone who knows its existence? For all we know, some non-wmf person might want to start hacking on the extension and want to view the test page (whatever it is).

Patch-set: 3
Reviewer: Brian Wolff <99@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0