Update patch set 3
Patch Set 3:

> I agree with Pmiazga, this should be permission based. I don't
 > think that it should use the "staff" group specificly (Since that
 > group is like wiki "root") but its easy to create random groups.

Is it? You have to put the i18n somewhere so either you coordinate with a random unrelated extension (WikimediaMessages) or put WMF-specific code in the extension or put a MessagesPreLoad or similar hook into the config. Each of those option seem ugly and way more complex than this.

 > Additionally, what's the point of hiding this? Is it a security
 > risk for normal users to find it (In which case, I'd like to know
 > details). If its not a security risk, what really is the problem
 > with having a temporary undocumented hard to discover test
 > interface that is open to anyone who knows its existence? For all
 > we know, some non-wmf person might want to start hacking on the
 > extension and want to view the test page (whatever it is).

Electron is known to die on large pages occasionally which halts all PDF rendering until manual intervention; exposing it to the book rendering feature might or might not make things worse than they already are, which is what the test page was supposed to test. Until we know for sure, I wanted to avoid some (possibly well-meaning) user testing some huge book and temporarily disabling book rendering for everyone.

Now that the book rendering feature has been delayed in favor of replacing Electron, I'm not really sure where to go with this.

Patch-set: 3