Update patch set 3
Patch Set 3:

Thanks for cleaning up behind yourself! However, this isn't really needed - the throttle exemption isn't active any more anyway (because it already expired) and there's no real need to remove it in a seperate patch (which has to be set up for a SWAT window to be merged and deployed). We usually just remove all outdated throttles when adding a new one (which removes the need for SWATing two patches, one for adding and one for removing). You may just abandon this and wait for somebody else to deal with it. Or, if you really want to deploy it, you can set it up for SWAT - add it to https://wikitech.wikimedia.org/wiki/Deployments and be around in the #wikimedia-operations IRC channel during the chosen SWAT window time.

---

Besides that, you seem to be stuck on the CI (jenkins-bot only gives you +1, not +2). That is because uploading code here and running tests on it means running that code (whatever it is) on Wikimedia servers. To prevent bad users from uploading malicious code, one needs to be on a whitelist in order to have all tests run (which will vote +2). If you're not on that whitelist, only a limited set of tests is run and you'll only receive +1.

As you're obviously not trying to do something bad, feel free to bug hashar to add you to that whitelist. ;)

Patch-set: 3
Reviewer: EddieGP <4565@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0