Update patch set 1
Patch Set 1: Code-Review+1

I might be missing some context, but Aude's suggestion seems reasonable if there is an intention to have the same setting on both labs and production. If this is only to be on Labs for now, the way it is submitted now seems fine.

Patch-set: 1
Label: Code-Review=+1