db-eqiad.php: Depool db1066
In order to to see if we can move db1066 to become vslow service in s3,
let's use db1080 to serve also API traffic as it was originally intended
on the gerrit example patchset for the hardware expansion.
Let's leave this config like this for a week or so and if all goes fine,
we can reallocate db1066 in s3 as vslow so we can decommission the old
host there and leave s3 ready.

Bug: T172679
Change-Id: Id086d1fac03f48e1a75ec81793cd1c6aa2565540