wikitech: Align 'contentadmin' and 'sysop' permissions
Three user rights are assigned to sysop's in some extension.json, which is
parsed after the configuration files are read. Hence those aren't
automatically added to contentadmins in the first line (assigning all
sysop permissions to contentadmin). Grant those explicitely. These are:
- Mass delete pages (nuke)
- Send a message to multiple users at once (massmessage)
- View the spam blacklist log (spamblacklistlog)

Also, the user right 'editrestrictedfield' is explicitely removed from
contentadmins. This used to be a right in use with PageForms, an
extension which has been undeployed from wikitech. Removing that line is
just cleanup.

Bug: T171208
Change-Id: I62e8a93760b639cdf0c60f13d117c74065030414