Update patch set 1
Patch Set 1: Code-Review-2

The Minerva config provides the logo as an image url, from /static, without a version hash. Normally unversioned urls are bad for cache, but these are fine because they're served from mediawiki-config/static, not mediawiki-core,skins,extensions - which are served via /w/static.php in production and should have a version hash appended by OutputPage::transformResourcePath, done automatically by CSSMin and ResourceLoader.

The /static served from mediawiki-config has an exception in Apache config to always cache it for long, even without a version hash, which means it might by default it won't roll over, but practice is to purge Varnish manually after deployment.

The code in Vector cannot use this logic in the current way because as being a file from outside MediaWiki, it cannot be read from disk, and has no url-to-file or file-to-url mapping (e.g. no wgScriptPath=$IP, or some such).

Using these in Vector as a url is fine, but if you intent to make use of ResourceLoader/CSSMin/embedding in some way, then they need to either be served from a MediaWiki skin/extension directory, or the Vector config needs to be given a local file path and external url separately so that it has a way to read the file and embed it, + fallback to normal url for debug mode and in case it is a non-SVG file (per CSSMin). The current method makes a lot of specific assumptions that break (e.g. large files embed the url "false", and any non-SVG file would break in IE8 due to the url not being known).

Patch-set: 1