Update patch set 4
Abandoned

The extension is not added to wikis anymore until its performance issues are fixed up. See T124841 which blocks this requests ( T177448 ).

We can easily restore this change once the extension can be enabled or that trwiktionary get allowed to enable it. That is better tracked on the tasks though :]

Patch-set: 4
Status: abandoned