Update patch set 6
Patch Set 6:

If not: 1. you go to the last jenkinsbot comment. 2. You'll see the Jenkins jobs done, and the report, you want to check each one with a FAILURE state (here only one: the -composer-hhvm-jessie) 3. click on the link to this job, you reach the console output 4. scroll down, check the errors (start by the one at the top, then see if the next one is another error or a consequence of the first).

Patch-set: 6
Reviewer: Dereckson <287@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0