Draft: Setting a multi-instance host
We are going to introduce the concept of multi-instance hosts in core.
Basically a host that will have two (or more) instances running on
different ports.

We would like know how to add it to mediawiki-config, and to confirm
that mediawiki-config supports "HOST:PORT" syntax

I have added a bunch of reviewers that I think that might help, but feel
free to remove yourself if you are not sure you can.

Bug: T178553
Change-Id: I0dc93e0feb5c11690c38dc74ee3e469944427f1d