Update patch set 3
Patch Set 3:

So this _should_ work, my concern is that there may be some code on mediawiki or outside (e.g. monitoring, TLS handling, etc.) that assumes db aliases are valid hostnames OR that assumes we will always be connecting to port 3306. This is relevant not only to multi-instance, but to proxying that some of the people here are about to work on.

To be fair, we could deploy this as is and wait for breakage on codfw as it is now depooled, but we should wait for feedback first.

Patch-set: 3
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0